feature: moved callbacks to from `app.py` to `callbacks.py`
